Phase 2 step 1: Bootstrap package skeleton + import-side-effect guard (no symbol moves)
