+++ apps/tui/io/__init__.py
@@ -1,0 +1,1 @@
+"""TUI I/O helpers (persistence, fetchers, stats, intraday)."""
