+++ apps/tui/render/__init__.py
@@ -1,0 +1,1 @@
+"""TUI render helpers (cells, text blocks, charts)."""
