+++ apps/tui/screens/__init__.py
@@ -1,0 +1,1 @@
+"""TUI modal screens."""
