+++ apps/tui/state/__init__.py
@@ -1,0 +1,1 @@
+"""TUI App state mixins package."""
