+++ apps/tui/state/mixins/__init__.py
@@ -1,0 +1,1 @@
+"""NepseDashboard method mixins."""
