+++ apps/tui/widgets/__init__.py
@@ -1,0 +1,1 @@
+"""TUI custom widgets."""
